backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
cryptography
argon2-cffi
pydantic
python-jose[cryptography]
passlib[bcrypt]
qrcode[pil]
faker 